requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
networkx>=2.8,<3.0
requests>=2.31.0
jinja2>=3.1.0
python-dotenv>=1.0.0
